+++ app/models/alchemy/page/page_natures.rb
@@ -7,6 +7,10 @@ module Alchemy
       already_public_for?(current_time) && still_public_for?(current_time)
     end
 
+    def expiration_time
+      public_until? ? public_until - Time.current : nil
+    end
+
     def taggable?
       definition['taggable'] == true
     end